Tenative implementation of testing for various fixed width integer types. Reviewing compiler flags. Tests passing for uint8. Too much repeated code.

+++ include/uint32.h
@@ -8,7 +8,7 @@ uint32_t
 fwuim_get_bit_uint32(uint8_t const bit_idx, uint32_t const vals)
 // returns the value (in lsb position) of the bit_idx-th bit in vals.
 {
-    return (vals >> bit_idx) & 0x1;
+    return (vals >> bit_idx) & (uint32_t)(0x1);
 }
 
 inline
@@ -16,7 +16,7 @@ uint32_t
 fwuim_set_bit_uint32(uint8_t const bit_idx, uint32_t const vals)
 // returns a value after setting the bit_idx-th bit in vals.
 {
-    return (0x1 << bit_idx) | vals;    
+    return ((uint32_t)(0x1) << bit_idx) | vals;    
 }
 
 inline
@@ -24,7 +24,7 @@ uint32_t
 fwuim_clear_bit_uint32(uint8_t const bit_idx, uint32_t const vals)
 // returns a value after clearing the bit_idx-th bit in vals.
 {
-    return (~(0x1 << bit_idx)) & vals;
+    return (~((uint32_t)(0x1) << bit_idx)) & vals;
 }
 
 inline
@@ -32,7 +32,7 @@ uint32_t
 fwuim_toggle_bit_uint32(uint8_t const bit_idx, uint32_t const vals)
 // returns a value after toggling the bit_idx-th bit in vals.
 {
-    return (0x1 << bit_idx) ^ vals;
+    return ((uint32_t)(0x1) << bit_idx) ^ vals;
 }
 
 #endif // FIXED_WIDTH_UNSIGNED_INTEGER_MANIPULATION_LIB_UINT32_H
